subsys: move config module into its own dir
Also this indents the proto:

According to the style guide:

    * Keep the line length to 80 characters.
    * Use an indent of 2 spaces.
