+++ mayastor/src/bin/initiator.rs
@@ -30,7 +30,7 @@ use mayastor::{
 };
 
 unsafe extern "C" fn run_static_initializers() {
-    spdk_sys::spdk_add_subsystem(subsys::MayastorSubsystem::new().0)
+    spdk_sys::spdk_add_subsystem(subsys::ConfigSubsystem::new().0)
 }
 
 #[used]
